kpopalbums/requirements.txt
google-cloud-bigquery
pandas
datetime
BeautifulSoup4
requests

